Batch 2: widen thread IDs to int64_t; clarify macOS stack-trace docs (#95, #97)
#95: macOS Mach thread IDs are 64-bit; the model stored ThreadInfo::tid as a
32-bit int, so values above INT_MAX truncated to negative (confirmed on real
hardware). Widen the identity end-to-end:
- ThreadInfo::tid and the details view-model's selected_thread_tid /
  cached_stack_tid -> int64_t.
- IProcessDataProvider::get_thread_stack(int pid, int tid) -> int64_t tid,
  across the interface and all six backend overrides (linux/procfs, windows,
  freebsd, macos, solaris). Other platforms' TIDs fit comfortably in int64_t,
  so this is a pure widening with no behavior change there.
- macOS get_threads: static_cast<int>(tids[i]) -> int64_t.
- Display: GUI/TUI print tid with %lld; GUI Threads sort uses a proper 3-way
  compare instead of subtraction (which could overflow once TIDs are large or
  negative). The other string-based sort comparators are batch 3 (#88).

#97: docs (README + PRIVILEGES.md) now state precisely why macOS thread stack
traces are unavailable — it's the hardened runtime/AMFI blocking task_for_pid
without a debugger entitlement, not merely SIP, and root does not lift it;
supporting it would require code-signing pex with a debugger entitlement.

Built GUI+TUI+tests on Linux; 60 checks pass. Cross-platform overrides
compiled by CI.

+++ src/core/model/process_info.hpp
@@ -59,7 +59,7 @@ struct ProcessInfo {
 };
 
 struct ThreadInfo {
-    int tid = 0;                    // Thread ID (Linux: TID, Windows: thread ID)
+    int64_t tid = 0;                // Thread ID (macOS Mach thread IDs are 64-bit; narrowing to int made them negative — issue #95)
     std::string name;               // Thread name if available
     char state = '?';               // Same state chars as ProcessInfo
     int priority = 0;
